style: improve the structure of app
